requirements.txt
PySide6>=6.5
Pillow>=9.1.0
mss>=9.0
pytesseract>=0.3.10
pyqtkeybind>=0.0.9
psutil>=5.9
pyinstaller>=6
pywin32>=306
requests>=2.31.0
